# fix: Use shared runtime for async commands (#26)