## Summary
Fixes #26 by implementing a shared Tokio runtime for async command execution instead of creating a new runtime for each async operation.

## Changes
- Added `Async` variant to `CmdInner` enum for proper async handling
- Modified `custom_async()` and `spawn()` to use shared runtime via `Cmd::async_cmd()`
- Updated `CommandExecutor` to properly handle async commands with the shared runtime
- Fixed future pinning issues with `Box::pin`

## Performance Impact
Created comprehensive benchmarks that prove the performance improvement:
- **Runtime creation overhead**: ~305µs per creation
- **Old approach**: 1.45ms per async operation (includes runtime creation)
- **New approach**: 1.13ms per async operation (shared runtime)
- **Improvement**: ~320µs saved per command (22% faster)

## Testing
- Added integration tests verifying async commands use shared runtime
- Added criterion benchmarks measuring actual performance improvement
- All existing tests pass

## Benchmark Results
```
running 6 benchmarks
create_new_runtime_per_operation: 1.45ms
use_shared_runtime: 1.13ms  
runtime_new_only: 305.37µs
spawn_on_shared_runtime: 1.12ms
old_approach_with_new_runtime: 1.45ms
new_approach_shared_runtime: 1.13ms
```

The benchmarks conclusively demonstrate the 22% performance improvement from eliminating runtime creation overhead.